Mobile rendering.
Goal: Fix mobile Rendering.

Important: Work in progress - contains dead, duplicate and in lined code.

Changes:

1. In actions.js, added a set_size action, that will set the current size to the redux state, making it available throughout the app.

2. In UnitSelect.js, changed the floating text to a prop.

3. In WeatherChart.js, moved the file to containers and connected it to the redux state. It now uses the current state and a function to compute a appropriate titleStyle.

4. In WeatherList.js, now using a function to compute the current size and setting the current size in the redux state.

5. In WeatherListNarrowScreen.js, removed some dead code, changed some styles and sending the name (for floating text ) to WeatherChart.

6. In WeatherListWideScreen.js, sending the name to WeatherChart.

7. In the reducers, added a reducer and connected the reducer for size.

8. In utils, added a array of custom breakpoints used to compute styles and a function to compute the current size.

9. Upgraded to MaterialUI 0.19.4. Appears to be mostly compatible.

Notes: A better and more elegant way of fixing this would be using a Grid System, as opposed to a custom breakpoint computation. However, a library ( Material UI ) specific Grid Layout System is not available in the current version. It is available in the beta version. However, the differences are to big, and would require a very big refactoring.